add a bunch more comments for diffing, scoring and why this scoring system is optimal
